requirements.txt
# EPI Recognition System - Requirements

# Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
pillow>=10.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Async
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Logging
colorlog>=6.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development (optional)
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Report Generation
reportlab>=4.0.0
openpyxl>=3.1.0

# Additional utilities
tqdm>=4.66.0
requests>=2.31.0
